refactor : ssd_driver.py 모듈 추출, 중복 함수 extract
