+++ shell.py
@@ -3,6 +3,7 @@ import subprocess
 import random
 import os
 import sys
+from ssd_driver import SSDDriver
 
 ROOT_DIR = os.path.dirname(__file__)
 
@@ -16,42 +17,6 @@ READ_ERROR = ERROR
 ERASE_SUCCESS = SUCCESS
 ERASE_ERROR = ERROR
 MAX_ERASE_SIZE = 10
-
-
-class SSDDriver:
-    def run_ssd_write(self, address: str, value: str):
-        command = ['python', 'ssd.py', 'W', str(address), str(value)]
-        result = subprocess.run(command, cwd=ROOT_DIR, capture_output=True, text=True)
-
-        if result.returncode == 0:
-            return WRITE_SUCCESS
-        else:
-            return WRITE_ERROR
-
-    def run_ssd_erase(self, address: str, lba_size: str):
-        command = ['python', 'ssd.py', 'E', str(address), str(lba_size)]
-        result = subprocess.run(command, cwd=ROOT_DIR, capture_output=True, text=True)
-
-        if result.returncode == 0:
-            return ERASE_SUCCESS
-        else:
-            return ERASE_ERROR
-
-    def run_ssd_read(self, address: str):
-        command = ['python', 'ssd.py', 'R', str(address)]
-        result = subprocess.run(command, cwd=ROOT_DIR, capture_output=True, text=True)
-        if result.returncode == 0:
-            return READ_SUCCESS
-        else:
-            return READ_ERROR
-
-    def get_ssd_output(self, file_path: str = ROOT_DIR + "/ssd_output.txt"):
-        with open(file_path, 'r', encoding='utf-8') as f:
-            line = f.readline().rstrip("\n")
-
-        if len(line) != 10:
-            raise ValueError(f"Error, value Length: {len(line)})")
-        return line
 
 
 class TestShellApp:
