+++ ssd_driver.py
@@ -1,0 +1,35 @@
+import subprocess
+import os
+
+ROOT_DIR = os.path.dirname(__file__)
+SUCCESS = 0
+ERROR = -1
+
+
+class SSDDriver:
+    def run_cmd_to_ssd(self, command):
+        result = subprocess.run(command, cwd=ROOT_DIR, capture_output=True, text=True)
+        if result.returncode == 0:
+            return SUCCESS
+        else:
+            return ERROR
+
+    def run_ssd_write(self, address: str, value: str):
+        command = ['python', 'ssd.py', 'W', str(address), str(value)]
+        return self.run_cmd_to_ssd(command)
+
+    def run_ssd_erase(self, address: str, lba_size: str):
+        command = ['python', 'ssd.py', 'E', str(address), str(lba_size)]
+        return self.run_cmd_to_ssd(command)
+
+    def run_ssd_read(self, address: str):
+        command = ['python', 'ssd.py', 'R', str(address)]
+        return self.run_cmd_to_ssd(command)
+
+    def get_ssd_output(self, file_path: str = ROOT_DIR + "/ssd_output.txt"):
+        with open(file_path, 'r', encoding='utf-8') as f:
+            line = f.readline().rstrip("\n")
+
+        if len(line) != 10:
+            raise ValueError(f"Error, value Length: {len(line)})")
+        return line
